+++ src/ApiTrait.php
@@ -30,22 +30,32 @@ trait ApiTrait
      * @param SystemStrand $strand The strand executing the API call.
      * @param mixed        $key    The yielded key.
      * @param mixed        $value  The yielded value.
+     *
+     * @return Generator|null
      */
     public function __dispatch(SystemStrand $strand, $key, $value)
     {
         if (null === $value) {
-            $this->cooperate($strand);
-        } elseif (\is_integer($value) || \is_float($value)) {
-            $this->sleep($strand, $value);
-        } elseif (\is_array($value)) {
-            $this->all($strand, ...$value);
-        } elseif (\is_resource($value)) {
+            return $this->cooperate($strand);
+        }
+
+        if (\is_integer($value) || \is_float($value)) {
+            return $this->sleep($strand, $value);
+        }
+
+        if (\is_array($value)) {
+            return $this->all($strand, ...$value);
+        }
+
+        if (\is_resource($value)) {
             if (\is_string($key)) {
-                $this->write($strand, $value, $key, PHP_INT_MAX);
+                return $this->write($strand, $value, $key, PHP_INT_MAX);
             } else {
-                $this->read($strand, $value, 1, PHP_INT_MAX);
+                return $this->read($strand, $value, 1, PHP_INT_MAX);
             }
-        } elseif (\method_exists($value, 'then')) {
+        }
+
+        if (\method_exists($value, 'then')) {
             $onFulfilled = static function ($result) use ($strand) {
                 $strand->send($result);
             };
@@ -69,17 +79,21 @@ trait ApiTrait
                     $value->cancel();
                 });
             }
-        } else {
-            $strand->throw(
-                new UnexpectedValueException(
-                    'The yielded pair ('
-                    . Repr::repr($key)
-                    . ', '
-                    . Repr::repr($value)
-                    . ') does not describe any known operation.'
-                )
-            );
+
+            return null;
         }
+
+        $strand->throw(
+            new UnexpectedValueException(
+                'The yielded pair ('
+                . Repr::repr($key)
+                . ', '
+                . Repr::repr($value)
+                . ') does not describe any known operation.'
+            )
+        );
+
+        return null;
     }
 
     /**
@@ -308,6 +322,14 @@ trait ApiTrait
      */
     public function all(SystemStrand $strand, ...$coroutines)
     {
+        if (empty($coroutines)) {
+            return (function () {
+                yield Recoil::cooperate();
+
+                return [];
+            })();
+        }
+
         $kernel = $strand->kernel();
         $substrands = [];
 
@@ -316,6 +338,8 @@ trait ApiTrait
         }
 
         (new StrandWaitAll(...$substrands))->await($strand);
+
+        return null;
     }
 
     /**
@@ -331,6 +355,10 @@ trait ApiTrait
      */
     public function any(SystemStrand $strand, ...$coroutines)
     {
+        if (empty($coroutines)) {
+            return $this->cooperate($strand);
+        }
+
         $kernel = $strand->kernel();
         $substrands = [];
 
@@ -339,6 +367,8 @@ trait ApiTrait
         }
 
         (new StrandWaitAny(...$substrands))->await($strand);
+
+        return null;
     }
 
     /**
@@ -357,18 +387,26 @@ trait ApiTrait
     {
         $max = \count($coroutines);
 
-        if ($count < 1 || $count > $max) {
+        if ($count < 0 || $count > $max) {
             $strand->throw(
                 new InvalidArgumentException(
                     'Can not wait for '
                     . $count
-                    . ' coroutines, count must be between 1 and '
+                    . ' coroutines, count must be between 0 and '
                     . $max
                     . ', inclusive.'
                 )
             );
 
-            return;
+            return null;
+        }
+
+        if ($count == 0 || empty($coroutines)) {
+            return (function () {
+                yield Recoil::cooperate();
+
+                return [];
+            })();
         }
 
         $kernel = $strand->kernel();
@@ -379,6 +417,8 @@ trait ApiTrait
         }
 
         (new StrandWaitSome($count, ...$substrands))->await($strand);
+
+        return null;
     }
 
     /**
@@ -394,6 +434,10 @@ trait ApiTrait
      */
     public function first(SystemStrand $strand, ...$coroutines)
     {
+        if (empty($coroutines)) {
+            return $this->cooperate($strand);
+        }
+
         $kernel = $strand->kernel();
         $substrands = [];
 
@@ -402,6 +446,8 @@ trait ApiTrait
         }
 
         (new StrandWaitFirst(...$substrands))->await($strand);
+
+        return null;
     }
 
     /**